renamed sources to modules, moved SourceThreads to threadmanagement module, in order to make output modules pluggable later
